+++ src/GafferCycles/CyclesOptions.cpp
@@ -163,20 +163,6 @@ CyclesOptions::CyclesOptions( const std::string &name )
 
 	// Dicing camera
 	options->addChild( new Gaffer::NameValuePlug( "cycles:dicing_camera", new IECore::StringData(), false, "dicingCamera" ) );
-
-	// Texture cache
-	options->addChild( new Gaffer::NameValuePlug( "cycles:texture:use_texture_cache", new IECore::BoolData( false ), false, "useTextureCache" ) );
-	options->addChild( new Gaffer::NameValuePlug( "cycles:texture:cache_size", new IECore::IntData( 1024 ), false, "textureCacheSize" ) );
-	options->addChild( new Gaffer::NameValuePlug( "cycles:texture:auto_convert", new IECore::BoolData( true ), false, "textureAutoConvert" ) );
-	options->addChild( new Gaffer::NameValuePlug( "cycles:texture:accept_unmipped", new IECore::BoolData( true ), false, "textureAcceptUnmipped" ) );
-	options->addChild( new Gaffer::NameValuePlug( "cycles:texture:accept_untiled", new IECore::BoolData( true ), false, "textureAcceptUntiled" ) );
-	options->addChild( new Gaffer::NameValuePlug( "cycles:texture:auto_tile", new IECore::BoolData( true ), false, "textureAutoTile" ) );
-	options->addChild( new Gaffer::NameValuePlug( "cycles:texture:auto_mip", new IECore::BoolData( true ), false, "textureAutoMip" ) );
-	options->addChild( new Gaffer::NameValuePlug( "cycles:texture:tile_size", new IECore::IntData( 64 ), false, "textureTileSize" ) );
-	options->addChild( new Gaffer::NameValuePlug( "cycles:texture:blur_diffuse", new IECore::FloatData( 0.0156f ), false, "textureBlurDiffuse" ) );
-	options->addChild( new Gaffer::NameValuePlug( "cycles:texture:blur_glossy", new IECore::FloatData( 0.0f ), false, "textureBlurGlossy" ) );
-	options->addChild( new Gaffer::NameValuePlug( "cycles:texture:use_custom_cache_path", new IECore::BoolData( false ), false, "useCustomCachePath" ) );
-	options->addChild( new Gaffer::NameValuePlug( "cycles:texture:custom_cache_path", new IECore::StringData(), false, "customCachePath" ) );
 }
 
 CyclesOptions::~CyclesOptions()